feat: Add prompt for carbon footprint estimation and include file extension in image upload
This commit adds a prompt for estimating the carbon footprint of the image content and includes the file extension in the image upload request. The prompt asks the user to provide the carbon footprint estimation in the form of a number and unit. This enhancement improves the functionality of the image upload feature and allows for more accurate carbon footprint calculations.