chore: Add expo-secure-store plugin to app.json and update eas.json
This commit adds the "expo-secure-store" plugin to the "app.json" file and updates the "eas.json" file to require a minimum version of "9.0.1" for the "eas-cli". The addition of the "expo-secure-store" plugin enhances the security of the application, while the update to the "eas-cli" ensures compatibility with the latest features and bug fixes.