refactor: Add getUser function to db.js
This commit adds the getUser function to the db.js file. The function retrieves user information based on the provided email. It is used in the /users/:token endpoint to fetch user info for a given token. This enhancement improves the functionality of the application by allowing users to retrieve their information.