refactor: Remove unnecessary code and update welcome layout
This commit removes unnecessary code from the welcome layout in the AnonymousLayout component. It also updates the layout to redirect to the /home page using the Redirect component. These changes improve the user experience by simplifying the code and automatically redirecting users to the home page upon visiting the welcome screen.