feat: Add handleSignOut prop to SidebarComponent
This commit adds the handleSignOut prop to the SidebarComponent in the home page. The handleSignOut prop is used to handle the sign out functionality when the user clicks on the "Log out" button in the sidebar. This enhancement improves the user experience by providing a way for users to easily log out of the application.