feat: Add lastLogin field to AuthContext
This commit adds a new field, lastLogin, to the AuthContext in the ctx.tsx file. The lastLogin field is used to store the timestamp of the user's last login. This enhancement improves the functionality of the application by providing a way to track and display the user's last login time.