feat: Add getPredictionVertexAIText function to AuthContext
This commit adds the getPredictionVertexAIText function to the AuthContext in the ctx.tsx file. The getPredictionVertexAIText function is used to fetch predictions based on a given text input. It sends a POST request to the server with the text input and retrieves the prediction data. This enhancement improves the functionality of the application by providing a way to generate predictions based on text input.