feat: Add getActivities endpoint for retrieving user activities
This commit adds a new endpoint, `/users/:id/activities`, to the server. The endpoint is used to retrieve the activities of a user based on their ID. It fetches the activities from the database and returns them in the response. This enhancement improves the functionality of the application by providing a way to view and manage user activities.