feat: Add deleteUser function to delete user by id
This commit adds a new function, deleteUser, to the database module. The function allows for the deletion of a user by their id. It first checks if the user exists, and if not, it returns an error. Then it deletes the user from the users table, as well as related data from the carbon_footprint and user_activities tables. This enhancement improves the functionality of the application by providing a way to delete user accounts.